% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{create.big.file.model.case1}
\alias{create.big.file.model.case1}
\title{Simulate data matching the article}
\usage{
create.big.file.model.case1(fileX = NULL, fileY = NULL, size.min = 1,
  p = 400, q = 500, chunk.size = 10000, ng = NULL)
}
\arguments{
\item{fileX}{filename of output X. If NULL then no file is written and the object is returned.}

\item{fileY}{filename of output Y. If NULL then no file is written and the object is returned.}

\item{size.min}{minimum size for output X and Y given in required Gb. Default 1.}

\item{p}{number of parameters for X}

\item{q}{number of parameters for Y}

\item{chunk.size}{size for chunks}
}
\value{
write an X and Y file of required size.
}
\description{
Produce a Big data file for replicaiton of the paper results.
}
