% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls_internal.R
\name{my.norm}
\alias{my.norm}
\title{Compute Norm}
\usage{
my.norm(x)
}
\arguments{
\item{x}{matrix of numeric variables}
}
\value{
Output will be a numeric matrix or vector.
}
\description{
Return normed vector
}
