% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.bigsgPLS}
\alias{predict.bigsgPLS}
\title{predict.sgpls}
\usage{
\method{predict}{bigsgPLS}(object, newX, ng = 1, comps = object$ncomp,
  da = FALSE, ...)
}
\arguments{
\item{object}{An object of class bigsgPLS}

\item{newX}{matrix or big.matrix object to make prediction on.}

\item{ng}{The number of chuncks used to read in the data and process using parallel computing.}

\item{comps}{A vector with the number of components to use in the PLS fit.}

\item{da}{Discriminant analysis argument to provide class estimates.}

\item{...}{Further arguments passed for methods.}
}
\description{
Predicted values based on (sparse or sparse group) PLS models. Regression coefficient and new predictions are given using the new observations.
}
\examples{

set.seed(1)
library(bigmemory)
n <- 15000
p <- 50
X = scale(matrix(rnorm(n*p), ncol = p, nrow = n))
y = X[,1:5] \%*\% 1:5 + rnorm(n)

X.bm <- as.big.matrix(X)
y.bm <- as.big.matrix(y)

library(doParallel)
registerDoParallel(cores = 2)
getDoParWorkers()
fit.PLS <- bigsgpls(X.bm, y.bm, case = 4, H = 4, ng = 10, keepX = rep(5,4))
pred.fit <- predict(fit.PLS, newX = X, ng = 1)
round(pred.fit$Beta,3)

}
