% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{create.big.file.model.case1}
\alias{create.big.file.model.case1}
\title{Simulate data matching the article}
\usage{
create.big.file.model.case1(fileX = "X.csv", fileY = "Y.csv", size.min,
  p = 400, q = 500, chunk.size = 10000)
}
\arguments{
\item{fileX}{filename of output X}

\item{fileY}{filename of output Y}

\item{size.min}{minimum size for output X file}

\item{p}{number of parameters X block}

\item{q}{number of parameters Y block}

\item{chunk.size}{size for chunks}
}
\value{
write an X and Y file of required size.
}
\description{
Produce a Big data file for replicaiton of the paper results.
}
